Create separate SpeedPlot class
